companion: validate companion options for non-standalone usage as well
this is breaking change since the validation may throw an error
