+++ packages/@uppy/companion/src/standalone/helper.js
@@ -144,43 +144,6 @@ const getConfigPath = () => {
 }
 
 /**
- * validates that the mandatory companion options are set.
- * If it is invalid, it will console an error of unset options and exits the process.
- * If it is valid, nothing happens.
- *
- * @param {object} config
- */
-exports.validateConfig = (config) => {
-  const mandatoryOptions = ['secret', 'filePath', 'server.host']
-  /** @type {string[]} */
-  const unspecified = []
-
-  mandatoryOptions.forEach((i) => {
-    const value = i.split('.').reduce((prev, curr) => prev[curr], config)
-
-    if (!value) unspecified.push(`"${i}"`)
-  })
-
-  // vaidate that all required config is specified
-  if (unspecified.length) {
-    console.error('\x1b[31m', 'Please specify the following options',
-      'to run companion as Standalone:\n', unspecified.join(',\n'), '\x1b[0m')
-    process.exit(1)
-  }
-
-  // validate that specified filePath is writeable/readable.
-  // TODO: consider moving this into the companion module itself.
-  try {
-    // @ts-ignore
-    fs.accessSync(`${config.filePath}`, fs.R_OK | fs.W_OK)
-  } catch (err) {
-    console.error('\x1b[31m', `No access to "${config.filePath}".`,
-      'Please ensure the directory exists and with read/write permissions.', '\x1b[0m')
-    process.exit(1)
-  }
-}
-
-/**
  *
  * @param {string} url
  */
