+++ packages/@uppy/companion/src/standalone/index.js
@@ -162,12 +162,20 @@ app.get('/', (req, res) => {
   res.send(helper.buildHelpfulStartupMessage(companionOptions))
 })
 
-// initialize companion
-helper.validateConfig(companionOptions)
+let companionApp
+try {
+  // initialize companion
+  companionApp = companion.app(companionOptions)
+} catch (error) {
+  console.error('\x1b[31m', error.message, '\x1b[0m')
+  process.exit(1)
+}
+
+// add companion to server middlewear
 if (process.env.COMPANION_PATH) {
-  app.use(process.env.COMPANION_PATH, companion.app(companionOptions))
+  app.use(process.env.COMPANION_PATH, companionApp)
 } else {
-  app.use(companion.app(companionOptions))
+  app.use(companionApp)
 }
 
 // WARNING: This route is added in order to validate your app with OneDrive.
